Added XML for GoalTimes
- added XML documents to have all the goal times of the levels in one place, instead of seperate scenes
- moved some scripts around

+++ Assets/Script/Objectives/MedalHandler.cs
@@ -15,7 +15,7 @@ public class MedalHandler {
     //how much medals you need to unlock the next world
     static int[] worldUnlocks = { 5, 9, 14, 19, 23 };
 
-    public static int CalcMedalType(float bronzeTime, float silverTime, float goldTime, float totalTime)
+    public static int CalcMedalType(double bronzeTime, double silverTime, double goldTime, double totalTime)
     { 
         int medalTypeNum = 0;
